Temporary support for new Google Maps URL format
New URLs can no longer be simply parsed into embed URLs, but the module will now
recognize them and attempt to convert to old embed URLs. In many cases this will
fail, which is why new option 'use_coordinates' has been added to module config.

If 'use_coordinates' is checked, embedded maps will use coordinates instead of a
query parameter consisting of place name. This is accurate, but the title of map
will also be in coordinates (not the name of the place in question).

As a long-term solution this module should be converted to use JS API instead.